Unit tests

+++ src/ru/sbt/lesson2/implementings/Triangle.java
@@ -3,9 +3,9 @@ package ru.sbt.lesson2.implementings;
 import ru.sbt.lesson2.interfaces.Figure;
 
 public class Triangle implements Figure {
-    private int a,b,c;
+    private double a,b,c;
 
-    public Triangle(int a, int b, int c) {
+    public Triangle(double a, double b, double c) {
         this.a = a;
         this.b = b;
         this.c = c;
